+++ src/Game/src/Inventory.h
@@ -16,12 +16,15 @@ private:
     std::map<Item, CookieNumber> m_Inventory;
     std::string last_item_added;
     CookieNumber last_item_added_amount;
+    CookieNumber _cookiesPerTap = 1;
 
 public:
     void addItem(Item& item, CookieNumber amountToAdd);
     void removeItem(Item& item, CookieNumber amountToRemove);
     CookieNumber getItemCount(Item &item);
     CookieNumber getLastItemAddedAmount();
+    CookieNumber getCookiesPerTap();
+    void incrementCookiesPerTap(CookieNumber amount);
     const std::string& getLastItemAdded();
     std::map<Item, CookieNumber>& getInventory() { return m_Inventory; };
 };